test: move tests to dedicated files
